+++ src/utilities/signtx.ts
@@ -1,5 +1,6 @@
 import WalletStore from "@/store/WalletStore";
-import { MyAlgoWalletSession } from "@algo-builder/web";
+import { MyAlgoWalletSession, WebMode } from "@algo-builder/web";
+import { JsonPayload, WalletMultisigMetadata } from "@algo-builder/web/build/algo-signer-types";
 import algosdk, { encodeObj } from "algosdk";
 import { toRaw } from "vue";
 import { convertBase64ToUnit8Array, convertToBase64 } from "./convert";
@@ -25,4 +26,42 @@ export async function signMultisigUsingMyAlgoWallet(
 	}
 	const outputBase64 = convertToBase64(encodeObj(myAlgoSignedTxn));
 	return { base64: outputBase64, json: myAlgoSignedTxn };
+}
+
+export async function signMultisigUsingAlgosigner(txnBase64: string,
+	multisigParams: WalletMultisigMetadata) {
+	const walletStore = WalletStore();
+	let signAlgoSigner = walletStore.webMode as WebMode;
+	let jsonObject = algosdk.decodeObj(
+		convertBase64ToUnit8Array(txnBase64)
+	) as algosdk.EncodedSignedTransaction;
+
+	let msig = algosdk.Transaction.from_obj_for_encoding(
+		jsonObject.txn
+	);
+	const bytes = algosdk.encodeObj(msig.get_obj_for_encoding());
+	const txnBase64Signing = convertToBase64(bytes); // base64 of the transaction without signature
+
+	let signedTxn = await signAlgoSigner.signTransaction([
+		{
+			txn: txnBase64Signing,
+			msig: multisigParams,
+		},
+	]);
+	const json = signedTxn[0] as JsonPayload;
+
+	let combineBlob = encodeObj(json.blob as any)
+	// we have multiple signatures 
+	if (jsonObject.msig?.subsig.findIndex(item => item.s?.length) !== -1) {
+		const blob1 = convertBase64ToUnit8Array(txnBase64);
+		const blob2 = convertBase64ToUnit8Array(json.blob as string);
+		combineBlob = algosdk.mergeMultisigTransactions([
+			blob1,
+			blob2,
+		]);
+	}
+
+	const outputBase64 = convertToBase64(combineBlob);
+	const newJson = algosdk.decodeSignedTransaction(combineBlob);
+	return { base64: outputBase64, json: newJson, }
 }
